pyproject.toml
[tool.poetry]
name = "lycosidae_src"
version = "0.1.0"
description = ""
authors = ["Loann Desboulets PhD"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
urllib3 = "^2.2.3"
numpy = "^2.1.3"
python-dotenv = "^1.0.1"
minio = "^7.2.12"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
